// lib-ts/browser/Types.ts
/**
 * Basic interface to keep track of element sizes.
 */
export interface Size {
    height: number;
    width: number;
}

export enum Color {
    Red = "red",
    Black = "black",
    None = ""
}

export enum Suit {
    Hearts = "hearts",
    Diamonds = "diamonds",
    Spades = "spades",
    Cloves = "cloves"
}